+++ src/Shape.java
@@ -1,9 +1,9 @@
 public abstract class Shape implements ShapeMethods { //Shape superclass
     //indicating the drawpoint which every shape has
-    int xCoordinate;
-    int yCoordinate;
+    double xCoordinate;
+    double yCoordinate;
 
-    public Shape(int posX, int posY){ //constructor
+    public Shape(double posX, double posY){ //constructor
         this.xCoordinate = posX;
         this.yCoordinate = posY;
     }
